(refactor) Move some component in share module
