+++ src/app/components/components.module.ts
@@ -4,8 +4,6 @@ import { SharedModule } from '../shared/shared.module';
 
 import { MenuComponent } from './menu/menu.component';
 import { MenuItemComponent } from './menu-item/menu-item.component';
-import { ToolbarComponent } from './toolbar/toolbar.component';
-import { ToolbarBackComponent } from './toolbar-back/toolbar-back.component';
 import { SearchListComponent } from './search-list/search-list.component';
 import { SearchBoxComponent } from './search-box/search-box.component';
 import { LaneService } from '../intervention-survey/shared/services/lane.service';
@@ -19,8 +17,6 @@ import {DialogsService} from './shared/dialogs.service';
   ],
   exports: [
     MenuComponent,
-    ToolbarComponent,
-    ToolbarBackComponent,
     SearchListComponent,
     SearchBoxComponent,
     YesNoDialogComponent
@@ -28,8 +24,6 @@ import {DialogsService} from './shared/dialogs.service';
   declarations: [
     MenuComponent,
     MenuItemComponent,
-    ToolbarComponent,
-    ToolbarBackComponent,
     SearchListComponent,
     SearchBoxComponent,
     FilterByPipe,
