+++ src/app/intervention-maps/intervention-maps.component.html
@@ -1,4 +1,4 @@
-<app-form-toolbar></app-form-toolbar>
+<app-toolbar></app-toolbar>
 <div class="content-under-toolbar">
   <igo-map-browser
   	igoMapContext
