+++ src/app/intervention-report/intervention-report.component.html
@@ -1,4 +1,4 @@
-<app-form-toolbar></app-form-toolbar>
+<app-toolbar></app-toolbar>
 <div class="content-under-toolbar">
   <md-card>
     <img src="" />
