+++ src/app/intervention-survey/intervention-survey.component.html
@@ -3,7 +3,7 @@
     <app-form-menu [items]="menuItems" (select)="select($event)" ></app-form-menu>
   </md-sidenav>
 
-  <app-form-toolbar-back [title]="'627 rapide-plat sud'" [mode]="mode" [sidenav]="sidenav"></app-form-toolbar-back>
+  <app-toolbar-back [title]="'627 rapide-plat sud'" [mode]="mode" [sidenav]="sidenav"></app-toolbar-back>
   <div class="content-under-toolbar">
     <div [ngSwitch]="selectedMenu">
       <div *ngSwitchCase="'waterSupply'">
