+++ src/app/prevention-survey/prevention-survey.component.html
@@ -1,4 +1,4 @@
-<app-form-toolbar-back [title]="'627 rapide-plat sud'"></app-form-toolbar-back>
+<app-toolbar-back [title]="'627 rapide-plat sud'"></app-toolbar-back>
 <div class="content-under-toolbar">
   <app-prevention-survey-question-slider></app-prevention-survey-question-slider>
 </div>
