+++ src/app/shared/components/toolbar/toolbar.component.styl
