+++ src/app/shared/services/window-ref.service.ts
